model

+++ laonog-modules/laonog-admin-service/src/main/java/com/laonog/admin/mapper/SysUserMapper.java
@@ -13,7 +13,7 @@ public interface SysUserMapper {
      * @param sysUserDO
      * @return
      */
-    Integer insertSysUser(SysUser sysUserDO);
+    Long insertSysUser(SysUser sysUserDO);
 
     /**
      * 删除
